refactor(toolkit): move ContentFolder under experimental and remove from factory
- Move `unique_toolkit/content/folder/` to `unique_toolkit/experimental/content_folder/`
- Drop `ContentFolder` from `UniqueServiceFactory` (no `content_folder()` method, not in `register_known_services`)
- Stop re-exporting `ContentFolder` from `unique_toolkit` and `unique_toolkit.content`
- Update docs, examples, and cross-references to the new import path
- Update Unreleased CHANGELOG entry to reflect the experimental location

Made-with: Cursor
