+++ unique_toolkit/docs/examples_from_docs/folder_create_multiple_abs_paths.py
@@ -1,7 +1,7 @@
 # %%
 from __future__ import annotations
 
-from unique_toolkit import ContentFolder
+from unique_toolkit.experimental.content_folder import ContentFolder
 
 folder_service = ContentFolder.from_settings()
