+++ unique_toolkit/docs/examples_from_docs/folder_create_verify_delete.py
@@ -1,7 +1,7 @@
 # %%
 from __future__ import annotations
 
-from unique_toolkit import ContentFolder
+from unique_toolkit.experimental.content_folder import ContentFolder
 
 # Absolute path from the KB root; change if this collides with real data.
 DEMO_FOLDER_PATH = "/EntangledToolkitDocsDemo"
