+++ unique_toolkit/unique_toolkit/content/__init__.py
@@ -1,7 +1,4 @@
 from .constants import DOMAIN_NAME as DOMAIN_NAME
-from .folder import (
-    ContentFolder as ContentFolder,
-)
 from .schemas import (
     Content as Content,
 )
