+++ unique_toolkit/unique_toolkit/experimental/content_folder/service.py
@@ -6,7 +6,7 @@ import unique_sdk
 
 from unique_toolkit._common.validate_required_values import validate_required_values
 from unique_toolkit.app.unique_settings import UniqueSettings
-from unique_toolkit.content.folder.functions import (
+from unique_toolkit.experimental.content_folder.functions import (
     add_folder_access,
     add_folder_access_async,
     create_folders,
@@ -17,13 +17,13 @@ from unique_toolkit.content.folder.functions import (
     remove_folder_access,
     remove_folder_access_async,
 )
-from unique_toolkit.content.folder.functions import (
+from unique_toolkit.experimental.content_folder.functions import (
     delete_folder as delete_folder_api,
 )
-from unique_toolkit.content.folder.functions import (
+from unique_toolkit.experimental.content_folder.functions import (
     delete_folder_async as delete_folder_async_api,
 )
-from unique_toolkit.content.folder.schemas import (
+from unique_toolkit.experimental.content_folder.schemas import (
     CreatedFolder,
     FolderDetail,
     FolderInfo,
@@ -39,7 +39,7 @@ class ContentFolder:
 
     **Scope id:** A knowledge-base folder is a *content scope*; its id is ``scopeId`` in
     the HTTP API. The toolkit uses ``scope_id`` for that same string (e.g. the ``id`` field
-    on :class:`~unique_toolkit.content.folder.schemas.CreatedFolder`).
+    on :class:`~unique_toolkit.experimental.content_folder.schemas.CreatedFolder`).
 
     **Creating folders** — pick *one* style per call (type checkers enforce this via overloads):
 
@@ -522,7 +522,7 @@ class ContentFolder:
         caveats).
 
         If ``scope_accesses`` is empty, no second ``add-access`` call is made; the returned
-        :class:`~unique_toolkit.content.folder.schemas.FolderDetail` is a minimal view
+        :class:`~unique_toolkit.experimental.content_folder.schemas.FolderDetail` is a minimal view
         (leaf id/name and, when ``private_to_creator`` is true, the creator grants only).
         """
         created = self._create_folder_impl(
